# BUG/PERF: Series.replace with dtype="category"