@phofl 2023-01-18
This broke the ci, opened https://github.com/pandas-dev/pandas/pull/50848 to revert if there is no quick-win here